import all field sub modules with a field top module
